timerfd: new timerfd API
kernel.org commits
5e05ad7d4e3b11f935998882b5d9c3b257137f1b timerfd: introduce a new hrtimer_forward_now() function
4d672e7ac79b5ec5cdc90e450823441e20464691 timerfd: new timerfd API
f79c343e2e5ba82b9661e7287a42fac596bf367a timerfd: un-break CONFIG_TIMERFD
46a56c5a02430f80ef73357aa1295875c1cef2a7 [PATCH] Fix timerfd breakage on avr32
9692bd9c140618e3f6a2848900aee96c9cd8a65c [PATCH] timerfd: fix remaining architectures
6783fe6253045bf5cbdf25e4be296f91437c5523 [MIPS] Wire up the timerfd_*() o32 system calls
0cb845ec3719c4cc13cedb3f86002c37c50aff0a [PARISC] remove old timerfd syscall
c7d16736b8759f45a49d318bf3a8ca7fed7b6962 MIPS: Compat: Use generic 32-bit wrapers for sys_timerfd_{g,s}ettime
45cc2b96f20fa27088a650587e5d9dc5fa5e32c0 [PATCH] fs/timerfd.c should #include <linux/syscalls.h>

+++ release/src-rt/linux/linux-2.6/kernel/hrtimer.c
@@ -286,7 +286,7 @@ EXPORT_SYMBOL_GPL(ktime_add_ns);
 /*
  * Divide a ktime value by a nanosecond value
  */
-unsigned long ktime_divns(const ktime_t kt, s64 div)
+u64 ktime_divns(const ktime_t kt, s64 div)
 {
 	u64 dclc, inc, dns;
 	int sft = 0;
@@ -301,7 +301,7 @@ unsigned long ktime_divns(const ktime_t kt, s64 div)
 	dclc >>= sft;
 	do_div(dclc, (unsigned long) div);
 
-	return (unsigned long) dclc;
+	return dclc;
 }
 #endif /* BITS_PER_LONG >= 64 */
 
@@ -632,10 +632,9 @@ void unlock_hrtimer_base(const struct hrtimer *timer, unsigned long *flags)
  * Forward the timer expiry so it will expire in the future.
  * Returns the number of overruns.
  */
-unsigned long
-hrtimer_forward(struct hrtimer *timer, ktime_t now, ktime_t interval)
+u64 hrtimer_forward(struct hrtimer *timer, ktime_t now, ktime_t interval)
 {
-	unsigned long orun = 1;
+	u64 orun = 1;
 	ktime_t delta;
 
 	delta = ktime_sub(now, timer->expires);
